Added some additional string utilities for file paths - level objects now carry their path and report their name

+++ Engine/Level.cpp
@@ -7,7 +7,7 @@ CLevel::CLevel()
 void CLevel::OnPlay()
 {
 	// Execute each Layers OnPlay method
-	for (int i = 0; i < m_layers.size(); i++)
+	for (size_t i = 0; i < m_layers.size(); i++)
 	{
 		m_layers[i]->OnPlay();
 	}
@@ -15,7 +15,7 @@ void CLevel::OnPlay()
 
 void CLevel::Destroy()
 {
-	for (int i = 0; i < m_layers.size(); i++)
+	for (size_t i = 0; i < m_layers.size(); i++)
 	{
 		m_layers[i]->Destroy();
 	}
@@ -24,7 +24,7 @@ void CLevel::Destroy()
 
 void CLevel::OnThink()
 {
-	for (int i = 0; i < m_layers.size(); i++)
+	for (size_t i = 0; i < m_layers.size(); i++)
 	{
 		m_layers[i]->OnThink();
 	}
@@ -32,7 +32,7 @@ void CLevel::OnThink()
 
 void CLevel::Draw()
 {
-	for (int i = 0; i < m_layers.size(); i++)
+	for (size_t i = 0; i < m_layers.size(); i++)
 	{
 		m_layers[i]->Draw();
 	}
